refactor: refactor node renderer files
